requirements.txt
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
